app/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
authlib==1.2.0
requests==2.28.2
pymongo==4.3.3
openai==0.27.6
schedule==1.2.0
tiktoken==0.4.0